controllers/inbody/requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
paddlepaddle==2.5.2
paddleocr==2.7.0.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78 